refactor(saves): extract saveService (PR-16)
Move createSave, restoreOverwrite, restoreFork, and deleteSave out of the fat
server/routes/saves.ts handlers into server/services/saveService.ts, mirroring
the GameCreationService convention (class + exported singleton). The service
throws coded errors (INVALID_SNAPSHOT, SAVE_NOT_FOUND,
UNSUPPORTED_SNAPSHOT_VERSION, INVALID_SNAPSHOT_COLLECTIONS, UNAUTHORIZED) and the
routes map them to the exact HTTP status + JSON bodies as before. Handler bodies
(including all [RESTORE] Step N console logs, deletion ordering, and fork
id-remapping) are verbatim moves. validateSnapshotCollections and
convertTimestamps become private module helpers in the service.

routes/saves.ts 629 -> 139 lines. Behavior-preserving; characterization + save/
load feature tests + route-manifest all green.
